refactor: move the pending offences implementation to base
This is the only way to avoid circular imports later when utilizing the pending offences to the core. The reason is because of the `base.py` which contains a lot of functionality on a single file, unseparated. Ideally, the file should be split into multiple files and the pending offences implementation would not be moved.
